moved logout placement, and changed about
