+++ api_keys.js
@@ -1,12 +1,12 @@
 module.exports = {
   mongoURI:
-    ‘mongodb+srv://dbUser:dbPassWord@cluster0.todug.mongodb.net/myFirstDatabase?retryWrites=true&w=majority’,
+    'mongodb+srv://dbUser:dbPassWord@cluster0.todug.mongodb.net/myFirstDatabase?retryWrites=true&w=majority',
   google: {
     clientID:
-      ‘372400979747-nuse2cggfa403e6ma5a56o491fdlditq.apps.googleusercontent.com’,
-    clientSecret: ‘I0LUouuX6T0smgNIdQwxnezR’,
+      '372400979747-nuse2cggfa403e6ma5a56o491fdlditq.apps.googleusercontent.com',
+    clientSecret: 'I0LUouuX6T0smgNIdQwxnezR',
   },
   session: {
-    cookieKey: ‘this_is_a_random_cookie_key’,
+    cookieKey: 'this_is_a_random_cookie_key',
   },
 };
