+++ client/components/NavbarContainer.jsx
@@ -27,12 +27,12 @@ const NavbarContainer = () => {
   const renderContent = (user) => {
     if (typeof user === 'number') {
       return (
-        <div>
+        <div align="right">
           <Nav>
             <Nav.Link
               href="/auth/logout"
-              className="btn float-right"
-              style={{ float: 'right' }}
+              // className="btn"
+              // style={{ float: 'right' }}
             >
               Logout
             </Nav.Link>
@@ -51,18 +51,15 @@ const NavbarContainer = () => {
     }
   };
 
-  // useEffect(() => {
 
-  // }, [])
-  // console.log('this is the func');
   getAuthedStatus();
   return (
     <div>
       <Navbar bg="primary" expand="lg" className="p-2" variant="dark">
-        <Navbar.Brand href="/landing">OneNumber</Navbar.Brand>
+        <Navbar.Brand href="/landing">About</Navbar.Brand>
         <Navbar.Toggle aria-controls="basic-navbar-nav" />
         <Navbar.Collapse id="basic-navbar-nav">
-          <div>{renderContent(user)}</div>
+          <div align="right" >{renderContent(user)}</div>
         </Navbar.Collapse>
       </Navbar>
     </div>
